fix(frontend): resolve react-doctor findings in auth-context
`make lint.frontend.react-doctor` flagged auth-context.tsx. Fixes per each
rule's canonical recipe:

- only-export-components: move the token helpers (setAuthTokens, clearAuthTokens,
  getStoredToken) into a new non-component module lib/auth-tokens.ts so the
  context file exports only AuthProvider + useAuth, restoring Fast Refresh.
- no-react19-deprecated-apis: replace useContext(AuthContext) with use().
- no-effect-chain: drop the token-watching effect; call fetchUser directly from
  login() and the mount effect, keeping the SSR-safe token=null initial render.

Editing these files pulled two pre-existing warnings into the diff scan, fixed
here too: memoize the provider value (jsx-no-constructed-context-values) and use
a typographic ellipsis in the callback spinner text.
